+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 
 ### 1.0.0
  First public release[index.html](..%2F0000%20OSS%2Fa-sit-plus.github.io%2Findex.html)
+ 
+## 2.0
+ * Proper BIT STRING
+ * BitSet (100% Kotlin BitSet implementation)